Add new Websocket Service Endpoint tests
- move helper functions into websocket_suite_test.go
- remove websocket_test.go
